requirements.txt
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
google-auth==2.28.0
google-auth-oauthlib==1.2.0
gspread==6.0.2
idna==3.6
oauthlib==3.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
StrEnum==0.4.15
urllib3==2.2.1